refactor: 타입 정리 및 함수 순서 정리
- 메인 컴포넌트 상단으로 이동
- 타입 분리 및 정리
