+++ src/components/pages/mypage/MyPageDefault.tsx
@@ -1,17 +1,6 @@
+import { IMyPageDefaultProp } from "@/types/mypage/mypage"
+
 import ReviewStateButton from "./ReviewStateButton"
-
-interface IMyPageDefaultProp {
-  dataFetchingKey: string
-  hasReview: boolean | undefined
-  onClick: (value: boolean) => void
-}
-
-const getDefaultText = (key: string, review: boolean | undefined): string => {
-  if (key === "myMeeting") return "신청한 모임이"
-  if (key === "myOwnMeeting") return "만든 모임이"
-  if (key === "myReview" && !!review) return "작성한 리뷰가"
-  return "작성 가능한 리뷰가"
-}
 
 const MyPageDefault = ({ dataFetchingKey, hasReview, onClick }: IMyPageDefaultProp) => {
   const isMyReview = dataFetchingKey === "myReview"
@@ -26,3 +15,10 @@ const MyPageDefault = ({ dataFetchingKey, hasReview, onClick }: IMyPageDefaultPr
 }
 
 export default MyPageDefault
+
+const getDefaultText = (key: string, review: boolean | undefined): string => {
+  if (key === "myMeeting") return "신청한 모임이"
+  if (key === "myOwnMeeting") return "만든 모임이"
+  if (key === "myReview" && !!review) return "작성한 리뷰가"
+  return "작성 가능한 리뷰가"
+}
